requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
bayesian-optimization==1.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
datasets==1.17.0
debugpy==1.5.1
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.4
emoji==1.6.3
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
fonttools==4.28.5
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
google-play-scraper==1.0.3
grpcio==1.43.0
huggingface-hub==0.4.0
hyperopt==0.2.7
idna==3.3
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
kili==2.103.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
msgpack==1.0.3
multidict==5.2.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.6.7
numpy==1.22.1
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
protobuf==3.19.3
pure-eval==0.2.1
py4j==0.10.9.3
pyaml==21.10.1
pyarrow==6.0.1
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
ray==1.9.2
redis==4.1.1
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scikit-optimize==0.9.0
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
stack-data==0.1.4
tabulate==0.8.9
tensorboardX==2.4.1
threadpoolctl==3.0.0
tokenizers==0.10.3
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.15.0
typeguard==2.13.3
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
websocket-client==1.2.3
wrapt==1.13.3
xxhash==2.0.2
yarl==1.7.2
zipp==3.7.0
